requirements.txt
# Kafka客户端工具依赖项
PySide6>=6.0.0
kafka-python>=2.0.0